requirements.txt
Django >= 2.0
python3-openid >= 3.0.8
requests-oauthlib >= 0.3.0
requests
pyjwt[crypto] >= 1.7
django-searchable-encrypted-fields >= 0.1
